rename files to avoid redundant imports
