+++ client/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Healthy Eating",
+  "name": "Healthy Eating App",
   "icons": [
     {
       "src": "favicon.ico",
