+++ client/src/pages/Home/HomeStyles.tsx
@@ -5,7 +5,6 @@ const HomeWrapper = styled.div`
   flex-direction: column;
   justify-content: space-around;
   align-items: center;
-  width: clamp(300px, 100%, 1000px);
   height: 100vh;
   padding: 1rem;
 `;
