+++ client/src/pages/Home/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Search } from "../../components/index";
-import HomeTitle from "./Title/Title";
+import HomeTitle from "./Title";
 import HomeWrapper from "./HomeStyles";
 
 const Home = () => {
